Simulate invoice - migrate off cycle code to common module
