+++ packages/server/customer-os-common-module/service/invoice_service.go
@@ -16,6 +16,7 @@ import (
 	"github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/log"
 	"github.com/pkg/errors"
+	"sort"
 	"time"
 )
 
@@ -29,6 +30,7 @@ type InvoiceService interface {
 	VoidInvoice(ctx context.Context, invoiceId, appSource string) error
 
 	FillCycleInvoice(ctx context.Context, tenant, contractId string, invoiceEntity *neo4jentity.InvoiceEntity, sliEntities neo4jentity.ServiceLineItemEntities) (*neo4jentity.InvoiceEntity, []*invoicepb.InvoiceLine, error)
+	FillOffCyclePrepaidInvoice(ctx context.Context, tenant, contractId string, invoiceEntity *neo4jentity.InvoiceEntity, sliEntities neo4jentity.ServiceLineItemEntities) (*neo4jentity.InvoiceEntity, []*invoicepb.InvoiceLine, error)
 }
 type invoiceService struct {
 	log      logger.Logger
@@ -481,6 +483,170 @@ func (h *invoiceService) FillCycleInvoice(ctx context.Context, tenant, contractI
 	return invoiceEntity, invoiceLines, nil
 }
 
+func (h *invoiceService) FillOffCyclePrepaidInvoice(ctx context.Context, tenant, contractId string, invoiceEntity *neo4jentity.InvoiceEntity, sliEntities neo4jentity.ServiceLineItemEntities) (*neo4jentity.InvoiceEntity, []*invoicepb.InvoiceLine, error) {
+	span, ctx := opentracing.StartSpanFromContext(ctx, "InvoiceService.FillOffCyclePrepaidInvoice")
+	defer span.Finish()
+	span.SetTag(tracing.SpanTagTenant, tenant)
+	span.SetTag(tracing.SpanTagEntityId, invoiceEntity.Id)
+	span.LogFields(log.String("contractId", contractId))
+
+	// filter out not applicable SLIs
+	referenceDate := invoiceEntity.PeriodStartDate
+	filteredSliEntities := neo4jentity.ServiceLineItemEntities{}
+	for _, sliEntity := range sliEntities {
+		// process only monthly, quarterly and annually SLIs
+		if sliEntity.Billed != neo4jenum.BilledTypeMonthly &&
+			sliEntity.Billed != neo4jenum.BilledTypeQuarterly &&
+			sliEntity.Billed != neo4jenum.BilledTypeAnnually &&
+			sliEntity.Billed != neo4jenum.BilledTypeOnce {
+			continue
+		}
+		// SLIs that started on or after reference date are not applicable
+		if sliEntity.StartedAt.After(referenceDate) || sliEntity.StartedAt.Equal(referenceDate) {
+			continue
+		}
+		// One time invoiced and cancelled SLIs are not applicable
+		if sliEntity.Billed == neo4jenum.BilledTypeOnce {
+			if sliEntity.Quantity <= 0 || sliEntity.Price <= 0 {
+				continue
+			}
+			if sliEntity.Canceled {
+				continue
+			}
+			ilDbNodeAndInvoiceId, err := h.services.Neo4jRepositories.InvoiceLineReadRepository.GetLatestInvoiceLineWithInvoiceIdByServiceLineItemParentId(ctx, tenant, sliEntity.ParentID)
+			if err != nil {
+				tracing.TraceErr(span, err)
+				h.log.Errorf("Error getting latest invoice line for sli parent id {%s}: {%s}", sliEntity.ParentID, err.Error())
+				return nil, nil, err
+			}
+			if ilDbNodeAndInvoiceId != nil {
+				continue
+			}
+		}
+		filteredSliEntities = append(filteredSliEntities, sliEntity)
+	}
+	// sort SLIs by startedAt
+	sort.Slice(filteredSliEntities, func(i, j int) bool {
+		return filteredSliEntities[i].StartedAt.Before(filteredSliEntities[j].StartedAt)
+	})
+	// group SLIs by parent id
+	sliByParentID := map[string][]neo4jentity.ServiceLineItemEntity{}
+	for _, sliEntity := range filteredSliEntities {
+		sliByParentID[sliEntity.ParentID] = append(sliByParentID[sliEntity.ParentID], sliEntity)
+	}
+
+	span.LogFields(log.Int("result - amount of SLIs to process", len(filteredSliEntities)))
+
+	amount, vat := float64(0), float64(0)
+	var invoiceLines []*invoicepb.InvoiceLine
+
+	proratedSliFound := false
+	// iterate SLIs by parent id
+	for parentId, slis := range sliByParentID {
+		// get latest SLI that is active on reference date
+		var sliEntityToInvoice *neo4jentity.ServiceLineItemEntity
+		for _, sliEntity := range slis {
+			if sliEntity.IsActiveAt(invoiceEntity.PeriodStartDate) {
+				sliEntityToInvoice = &sliEntity
+			}
+		}
+		// if no SLI is active on reference date, skip
+		if sliEntityToInvoice == nil {
+			span.LogFields(log.String("result - no active SLI for parent id", parentId))
+			continue
+		}
+		// get invoice line for latest invoiced SLI per parent
+		ilDbNodeAndInvoiceId, err := h.services.Neo4jRepositories.InvoiceLineReadRepository.GetLatestInvoiceLineWithInvoiceIdByServiceLineItemParentId(ctx, tenant, parentId)
+		if err != nil {
+			tracing.TraceErr(span, err)
+			h.log.Errorf("Error getting latest invoice line for sli parent id {%s}: {%s}", parentId, err.Error())
+			return nil, nil, err
+		}
+		finalSLIAmount, calculatedSLIVat := float64(0), float64(0)
+		if sliEntityToInvoice.Billed == neo4jenum.BilledTypeOnce {
+			quantity := sliEntityToInvoice.Quantity
+			if quantity <= 0 {
+				quantity = 1
+			}
+			finalSLIAmount = utils.TruncateFloat64(float64(quantity)*sliEntityToInvoice.Price, 2)
+			if finalSLIAmount <= 0 {
+				continue
+			}
+			calculatedSLIVat = utils.TruncateFloat64(finalSLIAmount*sliEntityToInvoice.VatRate/100, 2)
+		} else {
+			proratedInvoicedSLIAmount := float64(0)
+			if ilDbNodeAndInvoiceId != nil {
+				previousInvoiceDbNode, err := h.services.Neo4jRepositories.InvoiceReadRepository.GetInvoiceById(ctx, tenant, ilDbNodeAndInvoiceId.LinkedNodeId)
+				if err != nil {
+					tracing.TraceErr(span, err)
+					h.log.Errorf("Error getting invoice {%s}: {%s}", ilDbNodeAndInvoiceId.LinkedNodeId, err.Error())
+					return nil, nil, err
+				}
+				previousInvoiceEntity := mapper.MapDbNodeToInvoiceEntity(previousInvoiceDbNode)
+				// if previous invoice is for different cycle, charge full amount
+				if !previousInvoiceEntity.PeriodEndDate.Before(invoiceEntity.PeriodEndDate) {
+					// calculate already invoiced amount, prorated for the period
+					invoiceLineEntity := mapper.MapDbNodeToInvoiceLineEntity(ilDbNodeAndInvoiceId.Node)
+					calculatedInvoicedSLIAmountFor1Year := calculateSLIAmountForCycleInvoicing(invoiceLineEntity.Quantity, invoiceLineEntity.Price, invoiceLineEntity.BilledType, neo4jenum.BillingCycleAnnuallyBilling)
+					proratedInvoicedSLIAmount = prorateAnnualSLIAmount(sliEntityToInvoice.StartedAt, invoiceEntity.PeriodEndDate, calculatedInvoicedSLIAmountFor1Year)
+					proratedInvoicedSLIAmount = utils.TruncateFloat64(proratedInvoicedSLIAmount, 2)
+				}
+			}
+
+			calculatedSLIAmountFor1Year := calculateSLIAmountForCycleInvoicing(sliEntityToInvoice.Quantity, sliEntityToInvoice.Price, sliEntityToInvoice.Billed, neo4jenum.BillingCycleAnnuallyBilling)
+			proratedSLIAmount := prorateAnnualSLIAmount(sliEntityToInvoice.StartedAt, invoiceEntity.PeriodEndDate, calculatedSLIAmountFor1Year)
+			proratedSLIAmount = utils.TruncateFloat64(proratedSLIAmount, 2)
+			finalSLIAmount = proratedSLIAmount - proratedInvoicedSLIAmount
+			span.LogFields(log.Float64(fmt.Sprintf("result - final amount for SLI with parent id %s", parentId), finalSLIAmount))
+			if finalSLIAmount <= 0 {
+				continue
+			}
+			calculatedSLIVat = utils.TruncateFloat64(finalSLIAmount*sliEntityToInvoice.VatRate/100, 2)
+			proratedSliFound = true
+		}
+		amount += finalSLIAmount
+		vat += calculatedSLIVat
+		invoiceLine := invoicepb.InvoiceLine{
+			Name:                    sliEntityToInvoice.Name,
+			Price:                   utils.TruncateFloat64(calculatePriceForBilledType(sliEntityToInvoice.Price, sliEntityToInvoice.Billed, invoiceEntity.BillingCycle), 2),
+			Quantity:                sliEntityToInvoice.Quantity,
+			Amount:                  finalSLIAmount,
+			Total:                   finalSLIAmount + calculatedSLIVat,
+			Vat:                     calculatedSLIVat,
+			ServiceLineItemId:       sliEntityToInvoice.ID,
+			ServiceLineItemParentId: sliEntityToInvoice.ParentID,
+		}
+		switch sliEntityToInvoice.Billed {
+		case neo4jenum.BilledTypeMonthly:
+			invoiceLine.BilledType = commonpb.BilledType_MONTHLY_BILLED
+		case neo4jenum.BilledTypeQuarterly:
+			invoiceLine.BilledType = commonpb.BilledType_QUARTERLY_BILLED
+		case neo4jenum.BilledTypeAnnually:
+			invoiceLine.BilledType = commonpb.BilledType_ANNUALLY_BILLED
+		case neo4jenum.BilledTypeOnce:
+			invoiceLine.BilledType = commonpb.BilledType_ONCE_BILLED
+		}
+		invoiceLines = append(invoiceLines, &invoiceLine)
+	}
+
+	if !proratedSliFound && len(invoiceLines) > 0 {
+		// if no prorated SLI found, then invoice contains only once billed SLIs
+		// accept the invoice if today is monthly anniversary of the contract invoicing start date
+
+		// UPDATE: The rule is on hold, invoice will be issued even if contains only one time SLIs
+
+		//if !isMonthlyAnniversary(invoiceEntity.PeriodEndDate.AddDate(0, 0, 1)) {
+		//	invoiceLines = []*invoicepb.InvoiceLine{}
+		//}
+	}
+
+	invoiceEntity.Amount = amount
+	invoiceEntity.Vat = vat
+	invoiceEntity.TotalAmount = amount + vat
+
+	return invoiceEntity, invoiceLines, nil
+}
+
 func calculateInvoiceCycleEnd(start time.Time, cycle enum.BillingCycle) time.Time {
 	var end time.Time
 	switch cycle {
@@ -542,4 +708,15 @@ func calculatePriceForBilledType(price float64, billed neo4jenum.BilledType, cyc
 	}
 
 	return 0
+}
+
+func prorateAnnualSLIAmount(startDate, endDate time.Time, amount float64) float64 {
+	start := utils.ToDate(startDate)
+	end := utils.ToDate(endDate)
+	days := end.Sub(start).Hours() / 24
+	proratedAmount := amount * (days / 365)
+	if proratedAmount <= 0 {
+		return 0
+	}
+	return proratedAmount
 }
